requirements.txt
# ── ApplAI Python dependencies ────────────────────────────────────────────
# Pin minor versions; patch updates are expected to be safe.
# Run: pip install -r requirements.txt
# Security audit: pip-audit -r requirements.txt

# ── HTTP / Scraping ───────────────────────────────────────────────────────
httpx==0.27.*                  # async HTTP client (scrapers, Telegram, Notion)
beautifulsoup4==4.12.*         # HTML parsing
lxml==5.2.*                    # fast HTML/XML parser (bs4 backend)
playwright==1.44.*             # headless browser for JS-heavy boards (optional)

# ── LLM Clients ───────────────────────────────────────────────────────────
google-genai>=1.0.0,<2.0.0         # Gemini API (Tier 1 scoring) — new official SDK
groq>=0.9.0,<1.0.0                 # Groq API (Tier 1 scoring — alternative to Gemini)
anthropic>=0.40.0,<1.0.0           # Anthropic API (Tier 2 — CV + cover letter generation)

# ── Document Generation ───────────────────────────────────────────────────
python-docx==1.1.*             # Cover letter .docx generation
# LaTeX → PDF compilation is done via pdflatex subprocess (no Python package)

# ── Discord ──────────────────────────────────────────────────────────────
# No extra package needed — Discord API is called directly via httpx

# ── Notion ────────────────────────────────────────────────────────────────
notion-client==2.2.*           # Official Notion API client

# ── Utilities ────────────────────────────────────────────────────────────
python-dotenv==1.0.*           # Load .env files for local dev
tenacity==8.3.*                # Retry logic for scrapers and API calls
ratelimit==2.2.*               # Simple rate-limit decorator
python-dateutil==2.9.*         # Flexible date parsing from job boards
bleach==6.1.*                  # HTML sanitisation (extra layer in sanitize.py)

# ── Testing ───────────────────────────────────────────────────────────────
pytest==8.2.*
pytest-asyncio==0.23.*
pytest-cov==5.0.*

# ── Security auditing (dev only) ──────────────────────────────────────────
pip-audit==2.7.*
bandit==1.7.*
detect-secrets==1.4.*
